pyproject.toml
[tool.poetry]
name = "exel-parser-bot"
version = "0.1.0"
description = ""
authors = ["d"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
aiofiles = "^24.1.0"
alembic = "^1.14.0"
aiogram = "^3.14.0"
python-dotenv = "^1.0.1"
asyncpg = "^0.30.0"
postgres = "^4.0"
pandas = "^2.2.3"
openpyxl = "^3.1.5"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
